requirements.txt
streamlit
PyPDF2
numpy
sentence-transformers
google-generativeai